# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Collaboration Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workrooms:
    get:
      operationId: listWorkrooms
      tags: [workrooms]
      summary: List all workrooms
      responses:
        "200":
          description: List of workrooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workroom"
    post:
      operationId: createWorkroom
      tags: [workrooms]
      summary: Create a workroom from a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkroomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workroom"
  /workrooms/{id}:
    get:
      operationId: getWorkroom
      tags: [workrooms]
      summary: Get a workroom with stages and tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workroom detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkroomDetail"
        "404":
          description: Not found
    patch:
      operationId: updateWorkroom
      tags: [workrooms]
      summary: Update a workroom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkroomUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workroom"
    delete:
      operationId: deleteWorkroom
      tags: [workrooms]
      summary: Delete a workroom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Workroom Stages ────────────────────────────────────────────────────────
  /workrooms/{workroomId}/stages:
    get:
      operationId: listWorkroomStages
      tags: [stages]
      summary: List stages for a workroom
      parameters:
        - name: workroomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkroomStage"
  /workrooms/{workroomId}/stages/{stageId}:
    patch:
      operationId: updateWorkroomStage
      tags: [stages]
      summary: Update stage status (advance, approve gate, etc.)
      parameters:
        - name: workroomId
          in: path
          required: true
          schema:
            type: integer
        - name: stageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkroomStageUpdate"
      responses:
        "200":
          description: Updated stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkroomStage"

  # ── Workroom Tasks ─────────────────────────────────────────────────────────
  /workrooms/{workroomId}/tasks:
    get:
      operationId: listWorkroomTasks
      tags: [tasks]
      summary: List tasks for a workroom
      parameters:
        - name: workroomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkroomTask"
    post:
      operationId: createWorkroomTask
      tags: [tasks]
      summary: Create a task in a workroom
      parameters:
        - name: workroomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkroomTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkroomTask"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkroomTaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkroomTask"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Activity Log ──────────────────────────────────────────────────────────
  /workrooms/{workroomId}/activity:
    get:
      operationId: listWorkroomActivity
      tags: [activity]
      summary: Get activity log for a workroom
      parameters:
        - name: workroomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEntry"
    post:
      operationId: addWorkroomActivity
      tags: [activity]
      summary: Add an activity entry
      parameters:
        - name: workroomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityEntry"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Workroom:
      type: object
      required: [id, name, templateId, templateName, sector, status, currentStageName, progress, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        templateId:
          type: integer
        templateName:
          type: string
        sector:
          type: string
        status:
          type: string
          description: "active | paused | completed | blocked"
        currentStageName:
          type: string
        progress:
          type: integer
          description: "0-100 percent"
        objective:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    WorkroomDetail:
      type: object
      required: [id, name, templateId, templateName, sector, status, currentStageName, progress, stages, tasks, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        templateId:
          type: integer
        templateName:
          type: string
        sector:
          type: string
        status:
          type: string
        currentStageName:
          type: string
        progress:
          type: integer
        objective:
          type: ["string", "null"]
        stages:
          type: array
          items:
            $ref: "#/components/schemas/WorkroomStage"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/WorkroomTask"
        createdAt:
          type: string
        updatedAt:
          type: string
    WorkroomInput:
      type: object
      required: [name, templateId]
      properties:
        name:
          type: string
          minLength: 1
        templateId:
          type: integer
        objective:
          type: string
    WorkroomUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        objective:
          type: string

    # ── WorkroomStage ─────────────────────────────────────────────────────────
    WorkroomStage:
      type: object
      required: [id, workroomId, order, name, stageType, operandPattern, status, createdAt]
      properties:
        id:
          type: integer
        workroomId:
          type: integer
        order:
          type: integer
        name:
          type: string
        stageType:
          type: string
          description: "kerja | gate"
        operandPattern:
          type: string
          description: "sequential | multiclaw | openclaw"
        status:
          type: string
          description: "pending | active | awaiting_gate | approved | rejected | completed"
        gateDecision:
          type: ["string", "null"]
        gateNote:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    WorkroomStageUpdate:
      type: object
      properties:
        status:
          type: string
        gateDecision:
          type: string
        gateNote:
          type: string

    # ── WorkroomTask ──────────────────────────────────────────────────────────
    WorkroomTask:
      type: object
      required: [id, workroomId, stageId, title, priority, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        workroomId:
          type: integer
        stageId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        assigneeRole:
          type: ["string", "null"]
        agentId:
          type: ["integer", "null"]
        priority:
          type: string
          description: "low | medium | high"
        status:
          type: string
          description: "todo | doing | blocked | done"
        output:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    WorkroomTaskInput:
      type: object
      required: [stageId, title]
      properties:
        stageId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        assigneeRole:
          type: string
        agentId:
          type: integer
        priority:
          type: string
        status:
          type: string
        output:
          type: string
    WorkroomTaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assigneeRole:
          type: string
        agentId:
          type: integer
        priority:
          type: string
        status:
          type: string
        output:
          type: string

    # ── Activity ──────────────────────────────────────────────────────────────
    ActivityEntry:
      type: object
      required: [id, workroomId, eventType, description, createdAt]
      properties:
        id:
          type: integer
        workroomId:
          type: integer
        eventType:
          type: string
        description:
          type: string
        actor:
          type: ["string", "null"]
        createdAt:
          type: string
    ActivityEntryInput:
      type: object
      required: [eventType, description]
      properties:
        eventType:
          type: string
        description:
          type: string
        actor:
          type: string

    # ── Dashboard ─────────────────────────────────────────────────────────────
...