# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEntry {
  id: number;
  workroomId: number;
  eventType: string;
  description: string;
  /** @nullable */
  actor?: string | null;
  createdAt: string;
}