# lib/api-zod/src/generated/types/workroomDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkroomStage } from './workroomStage';
import type { WorkroomTask } from './workroomTask';

export interface WorkroomDetail {
  id: number;
  name: string;
  templateId: number;
  templateName: string;
  sector: string;
  status: string;
  currentStageName: string;
  progress: number;
  /** @nullable */
  objective?: string | null;
  stages: WorkroomStage[];
  tasks: WorkroomTask[];
  createdAt: string;
  updatedAt: string;
}