# lib/api-zod/src/generated/types/workroomTask.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkroomTask {
  id: number;
  workroomId: number;
  stageId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  assigneeRole?: string | null;
  /** @nullable */
  agentId?: number | null;
  /** low | medium | high */
  priority: string;
  /** todo | doing | blocked | done */
  status: string;
  /** @nullable */
  output?: string | null;
  createdAt: string;
  updatedAt: string;
}