# lib/api-zod/src/generated/types/workroomTaskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkroomTaskInput {
  stageId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  assigneeRole?: string;
  agentId?: number;
  priority?: string;
  status?: string;
  output?: string;
}