# lib/api-zod/src/generated/types/workroomUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Collaboration Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkroomUpdate {
  name?: string;
  status?: string;
  objective?: string;
}